+++ kong/tracing/instrumentation.lua
@@ -170,7 +170,7 @@ function _M.http_client()
   http.request_uri = wrap
 end
 
---- Regsiter vailable_types
+--- Register available_types
 -- functions in this list will be replaced with NOOP
 -- if tracing module is NOT enabled.
 for k, _ in pairs(_M) do
@@ -196,6 +196,11 @@ function _M.request(ctx)
                  and ctx.KONG_PROCESSING_START * 1e6
                   or time_ns()
 
+  local http_flavor = ngx.req.http_version()
+  if type(http_flavor) == "number" then
+    http_flavor = string.format("%.1f", http_flavor)
+  end
+
   local active_span = tracer.start_span(span_name, {
     span_kind = 2, -- server
     start_time_ns = start_time,
@@ -204,7 +209,7 @@ function _M.request(ctx)
       ["http.url"] = req_uri,
       ["http.host"] = host,
       ["http.scheme"] = scheme,
-      ["http.flavor"] = ngx.req.http_version(),
+      ["http.flavor"] = http_flavor,
       ["net.peer.ip"] = client.get_ip(),
     },
   })
@@ -261,7 +266,16 @@ do
   available_types.dns_query = true
 end
 
+
 -- runloop
+function _M.runloop_before_header_filter()
+  local root_span = ngx.ctx.KONG_SPANS and ngx.ctx.KONG_SPANS[1]
+  if root_span then
+    root_span:set_attribute("http.status_code", ngx.status)
+  end
+end
+
+
 function _M.runloop_log_before(ctx)
   -- add balancer
   _M.balancer(ctx)